+++ Firmware/include/setfuncs.h
@@ -16,6 +16,10 @@ void setVoltage();
 void setLightMeter();
 
 void toggleLidar(bool lidarStatus);
+
+void retryLidarInit();
+
+void clearLidarDisplay(const char *placeholder);
 // ---------------------
 
 #endif // SETFUNCS_H